Removing Modulefile in favor of metadata.json
As explained by Puppetlabs:
Throughout the Puppet 3.x series, deprecation warnings were issued for Modulefile. If you still have a Modulefile, it will be treated like any other text file in the root directory of the module. You will need to move any metadata contained in it to the metadata.json.